Add configuration for AppVeyor
