+++ appveyor.yml
@@ -1,0 +1,20 @@
+environment:
+  matrix:
+  - PYTHON: C:\\Python35-x64
+clone_depth: 1
+cache:
+- '%LOCALAPPDATA%\pip\Cache\http'
+- '%LOCALAPPDATA%\pip\Cache\wheels'
+build: off
+before_test:
+- cmd: >-
+    %PYTHON%\\python.exe -m venv venv
+    venv\\Scripts\\activate.bat
+    python -c "import sys; print(sys.executable)"
+    python -m pip install -r requirements.txt
+    python -m pip install -r requirements-devel.txt
+    python -m unittest discover -b -v -s "tests/integration/store" -t .
+after_test:
+- cmd: '%PYTHON%\\python.exe setup.py sdist'
+artifacts:
+- path: dist\*
