+++ tests/fixture_setup/__init__.py
@@ -1,0 +1,45 @@
+import socket
+
+if hasattr(socket, "AF_UNIX"):
+    from ._unix import FakeSnapd  # noqa
+from ._fixtures import (  # noqa
+    BzrRepo,
+    CleanEnvironment,
+    GitRepo,
+    HgRepo,
+    FakeAptBaseDependency,
+    FakeAptCache,
+    FakeAptCachePackage,
+    FakeBaseEnvironment,
+    FakeElf,
+    FakeFilesystem,
+    FakeLXD,
+    FakeMetadataExtractor,
+    FakeProjectOptions,
+    FakeParts,
+    FakePartsServerRunning,
+    FakePartsWiki,
+    FakePartsWikiWithSlashes,
+    FakePartsWikiWithSlashesRunning,
+    FakePartsWikiOrigin,
+    FakePartsWikiOriginRunning,
+    FakePartsWikiRunning,
+    FakePlugin,
+    FakeSnapcraftctl,
+    FakeSnapcraftIsASnap,
+    FakeSSOServerRunning,
+    FakeStore,
+    FakeStoreAPIServerRunning,
+    FakeStoreSearchServerRunning,
+    FakeStoreUploadServerRunning,
+    FakeTerminal,
+    SharedCache,
+    SilentSnapProgress,
+    SnapcraftYaml,
+    StagingStore,
+    SvnRepo,
+    TempCWD,
+    TempXDG,
+    TestStore,
+    WithoutSnapInstalled,
+)
