+++ tests/fixture_setup/_fixtures.py
@@ -22,7 +22,6 @@ import os
 import platform
 import pkgutil
 import shutil
-import socketserver
 import string
 import subprocess
 import sys
@@ -47,7 +46,6 @@ from tests import fake_servers
 from tests.fake_servers import (
     api,
     search,
-    snapd,
     upload
 )
 from tests.subprocess_utils import (
@@ -1050,76 +1048,6 @@ class SnapcraftYaml(fixtures.Fixture):
         with open(os.path.join(self.path, 'snap', 'snapcraft.yaml'),
                   'w') as snapcraft_yaml_file:
             yaml.dump(self.data, snapcraft_yaml_file)
-
-
-class UnixHTTPServer(socketserver.UnixStreamServer):
-
-    def get_request(self):
-        request, client_address = self.socket.accept()
-        # BaseHTTPRequestHandler expects a tuple with the client address at
-        # index 0, so we fake one
-        if len(client_address) == 0:
-            client_address = (self.server_address,)
-        return (request, client_address)
-
-
-class FakeSnapd(fixtures.Fixture):
-
-    @property
-    def snaps_result(self):
-        self.request_handler.snaps_result
-
-    @snaps_result.setter
-    def snaps_result(self, value):
-        self.request_handler.snaps_result = value
-
-    @property
-    def snap_details_func(self):
-        self.request_handler.snap_details_func
-
-    @snap_details_func.setter
-    def snap_details_func(self, value):
-        self.request_handler.snap_details_func = value
-
-    @property
-    def find_result(self):
-        self.request_handler.find_result
-
-    @find_result.setter
-    def find_result(self, value):
-        self.request_handler.find_result = value
-
-    def __init__(self):
-        super().__init__()
-        self.request_handler = snapd.FakeSnapdRequestHandler
-        self.snaps_result = []
-        self.find_result = []
-        self.snap_details_func = None
-
-    def setUp(self):
-        super().setUp()
-        snapd_fake_socket_path = tempfile.mkstemp()[1]
-        os.unlink(snapd_fake_socket_path)
-
-        socket_path_patcher = mock.patch(
-            'snapcraft.internal.repo.snaps.get_snapd_socket_path_template')
-        mock_socket_path = socket_path_patcher.start()
-        mock_socket_path.return_value = 'http+unix://{}/v2/{{}}'.format(
-            snapd_fake_socket_path.replace('/', '%2F'))
-        self.addCleanup(socket_path_patcher.stop)
-
-        self._start_fake_server(snapd_fake_socket_path)
-
-    def _start_fake_server(self, socket):
-        self.server = UnixHTTPServer(socket, self.request_handler)
-        server_thread = threading.Thread(target=self.server.serve_forever)
-        server_thread.start()
-        self.addCleanup(self._stop_fake_server, server_thread)
-
-    def _stop_fake_server(self, thread):
-        self.server.shutdown()
-        self.server.socket.close()
-        thread.join()
 
 
 class SharedCache(fixtures.Fixture):
